# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IC7 Enterprise RAG Agent",
    "description": "\nAgente RAG empresarial para el sistema de **gestión de solicitudes internas**.\n\n* **/v1/chat** — preguntas en lenguaje natural; el agente busca en la documentación\n  (RAG híbrido) y usa herramientas (estado, prioridad, esfuerzo, resumen, servicios cloud).\n* **/v1/documents** — carga y procesamiento de documentos (PDF, DOCX, MD, TXT).\n* **/v1/history** — trazabilidad completa de interacciones.\n\nAutenticación: cabecera `X-API-Key` (deshabilitada solo en `ENVIRONMENT=local`).\n",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "operación"
        ],
        "summary": "Liveness",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "tags": [
          "operación"
        ],
        "summary": "Readiness",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/chat": {
      "post": {
        "tags": [
          "asistente"
        ],
        "summary": "Consultar al asistente",
        "description": "Envía una pregunta al agente. Devuelve respuesta, fuentes citadas, herramientas usadas,\nevaluación de fundamentación y banderas de seguridad.",
        "operationId": "chat_v1_chat_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents": {
      "post": {
        "tags": [
          "documentos"
        ],
        "summary": "Cargar y procesar documentos",
        "description": "Ingesta: extracción de texto -> chunking -> embeddings -> vector store.\nRe-cargar un archivo con el mismo nombre reemplaza su versión anterior.",
        "operationId": "upload_documents_v1_documents_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_v1_documents_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "documentos"
        ],
        "summary": "Listar documentos indexados",
        "operationId": "list_documents_v1_documents_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentInfo"
                  },
                  "title": "Response List Documents V1 Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/reindex": {
      "post": {
        "tags": [
          "documentos"
        ],
        "summary": "Re-indexar el corpus base",
        "operationId": "reindex_v1_documents_reindex_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{source}": {
      "delete": {
        "tags": [
          "documentos"
        ],
        "summary": "Eliminar un documento del índice",
        "operationId": "delete_document_v1_documents__source__delete",
        "parameters": [
          {
            "name": "source",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/history": {
      "get": {
        "tags": [
          "trazabilidad"
        ],
        "summary": "Consultar historial de interacciones",
        "operationId": "list_history_v1_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[A-Za-z0-9_-]{6,64}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(answered|no_info|blocked|output_blocked|incomplete)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/history/stats": {
      "get": {
        "tags": [
          "trazabilidad"
        ],
        "summary": "Métricas agregadas del historial",
        "operationId": "history_stats_v1_history_stats_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/history/{interaction_id}": {
      "get": {
        "tags": [
          "trazabilidad"
        ],
        "summary": "Detalle de una interacción",
        "operationId": "get_interaction_v1_history__interaction_id__get",
        "parameters": [
          {
            "name": "interaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interaction Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools": {
      "get": {
        "tags": [
          "asistente"
        ],
        "summary": "Herramientas disponibles para el agente",
        "operationId": "list_tools_v1_tools_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_documents_v1_documents_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "Uno o más archivos .pdf, .docx, .md, .txt"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_upload_documents_v1_documents_post"
      },
      "ChatRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Question",
            "examples": [
              "¿Cuál es el SLA de resolución de una solicitud P2?"
            ]
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]{6,64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Permite conversación multi-turno. Si se omite se crea una nueva sesión."
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "interaction_id": {
            "type": "string",
            "title": "Interaction Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "answered | no_info | blocked | output_blocked | incomplete"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolCallTrace"
            },
            "type": "array",
            "title": "Tool Calls"
          },
          "grounding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Grounding"
              },
              {
                "type": "null"
              }
            ]
          },
          "security": {
            "additionalProperties": true,
            "type": "object",
            "title": "Security"
          },
          "usage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Usage"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "interaction_id",
          "session_id",
          "answer",
          "status",
          "sources",
          "tool_calls",
          "grounding",
          "security",
          "usage",
          "model",
          "prompt_version",
          "latency_ms"
        ],
        "title": "ChatResponse"
      },
      "DocumentInfo": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "chunks": {
            "type": "integer",
            "title": "Chunks"
          }
        },
        "type": "object",
        "required": [
          "source",
          "chunks"
        ],
        "title": "DocumentInfo"
      },
      "Grounding": {
        "properties": {
          "score": {
            "type": "number",
            "title": "Score"
          },
          "grounded": {
            "type": "boolean",
            "title": "Grounded"
          },
          "evaluated_sentences": {
            "type": "integer",
            "title": "Evaluated Sentences"
          },
          "unsupported_sentences": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unsupported Sentences"
          },
          "invalid_citations": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Invalid Citations"
          }
        },
        "type": "object",
        "required": [
          "score",
          "grounded",
          "evaluated_sentences",
          "unsupported_sentences",
          "invalid_citations"
        ],
        "title": "Grounding"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryPage": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "items",
          "count"
        ],
        "title": "HistoryPage"
      },
      "IngestionResponse": {
        "properties": {
          "ingested": {
            "items": {
              "$ref": "#/components/schemas/IngestionResult"
            },
            "type": "array",
            "title": "Ingested"
          },
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          },
          "total_chunks_in_index": {
            "type": "integer",
            "title": "Total Chunks In Index"
          }
        },
        "type": "object",
        "required": [
          "ingested",
          "total_chunks_in_index"
        ],
        "title": "IngestionResponse"
      },
      "IngestionResult": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "chunks": {
            "type": "integer",
            "title": "Chunks"
          },
          "replaced_chunks": {
            "type": "integer",
            "title": "Replaced Chunks"
          },
          "flagged_chunks": {
            "type": "integer",
            "title": "Flagged Chunks"
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          },
          "elapsed_ms": {
            "type": "integer",
            "title": "Elapsed Ms"
          }
        },
        "type": "object",
        "required": [
          "source",
          "chunks",
          "replaced_chunks",
          "flagged_chunks",
          "embedding_model",
          "elapsed_ms"
        ],
        "title": "IngestionResult"
      },
      "Source": {
        "properties": {
          "ref": {
            "type": "integer",
            "title": "Ref"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "cited": {
            "type": "boolean",
            "title": "Cited"
          }
        },
        "type": "object",
        "required": [
          "ref",
          "source",
          "section",
          "score",
          "snippet",
          "cited"
        ],
        "title": "Source"
      },
      "ToolCallTrace": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "elapsed_ms": {
            "type": "integer",
            "title": "Elapsed Ms"
          }
        },
        "type": "object",
        "required": [
          "name",
          "arguments",
          "ok",
          "elapsed_ms"
        ],
        "title": "ToolCallTrace"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}